Updated python 3 script

+++ minifigs/trainMinifigNetwork.py
@@ -46,11 +46,14 @@ for folder_idx, folder in enumerate(names):
     for idx, file in enumerate(files):
         file_name = BASE_DIR + folder + file
         if idx < n_train:
-            shutil.move(file_name, BASE_DIR + "train/" + names[folder_idx])
+            if os.path.isfile(file_name, BASE_DIR):
+                shutil.move(file_name, BASE_DIR + "train/" + names[folder_idx])
         elif idx < n_train + n_valid:
-            shutil.move(file_name, BASE_DIR + "val/" + names[folder_idx])
+            if os.path.isfile(file_name, BASE_DIR):
+                shutil.move(file_name, BASE_DIR + "val/" + names[folder_idx])
         else:
-            shutil.move(file_name, BASE_DIR + "test/" + names[folder_idx])
+            if os.path.isfile(file_name, BASE_DIR):
+                shutil.move(file_name, BASE_DIR + "test/" + names[folder_idx])
 
 
 # Generate batches of tensor image data with
